+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Unreleased ###
+* WooCommerce webhook trigger "order.paid"
 
 ### 1.0.5: May 14nd, 2018
 * Fix PDF title if order number is empty